4.13-4.22

+++ part4/blogs/controllers/blogs.js
@@ -1,8 +1,13 @@
 const blogsRouter = require("express").Router()
 const Blog = require("../models/blog")
+const { userExtractor } = require("../utils/middleware")
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate("user", {
+    username: 1,
+    name: 1,
+    id: 1,
+  })
   response.json(blogs)
 })
 
@@ -15,44 +20,73 @@ blogsRouter.get("/:id", async (request, response) => {
   }
 })
 
-blogsRouter.post("/", async (request, response) => {
-  const body = request.body
+blogsRouter.post("/", userExtractor, async (request, response) => {
+  const { user } = request
+
+  if (!user) {
+    return response.status(401).json({ error: "token invalid" })
+  }
+
+  const { title, author, url, likes } = request.body
+
+  if (!title || !url) {
+    return response.status(400).json({ error: "title or url is missing" })
+  }
 
   const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url || "",
-    likes: body.likes || 0,
+    title,
+    author,
+    url,
+    likes: likes ?? 0,
+    user: user,
   })
 
   const savedBlog = await blog.save()
+
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save({ validateModifiedOnly: true })
+
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete("/:id", userExtractor, async (request, response) => {
+  const { user } = request
+
+  if (!user) {
+    return response.status(401).json({ error: "token invalid" })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog?.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: "unauthorized access" })
+  }
 })
 
-blogsRouter.put("/:id", (request, response, next) => {
-  const { title, author, url, likes } = request.body
+blogsRouter.put("/:id", async (request, response, next) => {
+  try {
+    const { title, author, url, likes } = request.body
 
-  Blog.findById(request.params.id)
-    .then((blog) => {
-      if (!blog) {
-        return response.status(404).end()
-      }
+    const blog = await Blog.findById(request.params.id)
 
-      blog.title = title
-      blog.author = author
-      blog.url = url
-      blog.likes = likes
+    if (!blog) {
+      return response.status(404).end()
+    }
 
-      return blog.save().then((updatedBlog) => {
-        response.json(updatedBlog)
-      })
-    })
-    .catch((error) => next(error))
+    blog.title = title
+    blog.author = author
+    blog.url = url
+    blog.likes = likes
+
+    const updatedBlog = await blog.save()
+    return response.json(updatedBlog)
+  } catch (error) {
+    next(error)
+  }
 })
 
 module.exports = blogsRouter
